Move Book admin functions to Admin::BooksController. Create and remove views as necessary
