+++ app/views/admin/books/_filters.html.erb
@@ -1,0 +1,15 @@
+<div class="row store-top-bar">
+  <div class="col-md-12 store-order-field">
+    <%= form_tag '/', method: :get do %>
+      <%= label_tag "search" %>
+      <%= text_field_tag "search" %>
+
+      <%= label_tag "sort_field", "Order books by: "%>
+      <%= select_tag "sort_field", options_for_select(@sortable_fields) %>
+      <%= select_tag "sort_order",
+        options_for_select([['Ascending', 'ASC'],['Descending', 'DESC']]) %>
+
+      <%= submit_tag "Submit", class: "btn btn-primary" %>
+    <% end %>
+  </div>
+</div>
