+++ app/views/admin/books/index.html.erb
@@ -1,0 +1,22 @@
+<%= render 'filters' %>
+
+<div class="row store-listing-header">
+  <div class="col-md-3">Title</div>
+  <div class="col-md-2">Author</div>
+  <div class="col-md-2">Published Date</div>
+  <div class="col-md-1">Price</div>
+  <div class="col-md-2">Category</div>
+</div>
+
+<div class="store-book-list">
+  <% @books.each do |book| %>
+    <div class="row">
+      <div class="col-md-3 store-book-title"><%= link_to book.title, book_path(book) %></div>
+      <div class="col-md-2 store-book-author"><%= book.author %></div>
+      <div class="col-md-2 store-book-published-date"><%= book.published_date %></div>
+      <div class="col-md-1 store-book-price">$<%= book.price %></div>
+      <div class="col-md-2 store-book-category"><%= book.category %></div>
+    </div>
+  <% end %>
+</div>
+<%= paginate @books %>
