+++ app/views/admin/dashboard/index.html.erb
@@ -1,0 +1,3 @@
+<p><%= link_to 'Create a new admin', '#' %></p>
+<p><%= link_to 'List a new book for sale', new_admin_book_path %></p>
+<p><%= link_to 'View a list of orders', '#' %></p>
