+++ app/views/layouts/application.html.erb
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html>
   <head>
-    <title>The Ugly Bookstore</title>
+    <title>The Beautiful Rails Bookstore</title>
     <%= stylesheet_link_tag    'application', media: 'all'%>
     <%= javascript_include_tag 'application'%>
     <%= csrf_meta_tags %>
